Better macro expansion
The macro expansion now is not just copying macro body into input file.
Now when we encounter a call to a macro, we simply change the state of
the tokenizer to the macro body source. It allows us to process all
statements in the macro body like when we don't call to a macro